# web/backend/static/openapi.yaml
openapi: 3.1.0
info:
  title: ESA Host Brief Viewer
  version: 0.2.0
  description: 'ESA Incentive Pipeline backend API.


    **Authentication**


    - Host endpoints: require `esa_session` cookie (set via `POST /api/login`). Valid
    24h.

    - Admin endpoints: require `esa_admin_session` cookie (set via `POST /api/admin/login`).
    Valid 1h.


    **Base URL**: `https://esa-hosting.codescales.xyz`


    **AI/Tool discovery**


    - OpenAPI spec (this file): `GET /openapi.yaml`

    - Natural-language context doc: `GET /llms.txt`

    - Interactive UI: `GET /docs`'
  contact:
    name: ESA Incentive Pipeline
    url: https://github.com/BongoEADGC6/esa-incentive-pipeline
paths:
...
  /api/runs:
    get:
      tags:
      - runs
      summary: List Runs
      operationId: list_runs_api_runs_get
      parameters:
      - name: stream
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Stream
      - name: window
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Window
      - name: next_hours
        in: query
        required: false
        schema:
          type: number
          default: 0
          title: Next Hours
      - name: marathon
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Marathon
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runs/{slug}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: get_run_api_runs__slug__get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/runners/{slug}:
    get:
      tags:
      - runners
      summary: Get Runner
      operationId: get_runner_api_runners__slug__get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerDTOResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: 'Returns runner identity and **current-marathon** run counts only.


        - `run_count` — number of runs in this event''s schedule

        - `esa_count` — distinct streams in this event (not all-time event count)

        - `first_esa` — earliest scheduled run date in the current marathon


        For all-time historical data (17 events, 26 appearances, first debut year,
        country, SRC tenure, ESA run history) call **GET /api/runners/{slug}/profile**.'
  /api/runners/{slug}/profile:
    get:
      tags:
      - runners
      summary: Get Runner Profile
      description: 'Returns the full composite runner profile sourced from lineup.esamarathon.dev
        and speedrun.com.


        This is the authoritative source for all historical ESA data:


        - `summary.event_count` — all-time ESA events attended

        - `summary.appearance_count` — all-time ESA run appearances

        - `summary.first_event` / `first_event_year` — historical debut

        - `summary.esa_game_count` — distinct games run at ESA across all events

        - `summary.country_name` — self-reported, unverified

        - `summary.src_tenure_years` — years on speedrun.com

        - `stats.past_esa.events[]` — ordered list of all event names attended

        - `stats.past_esa.esa_runs[]` — per-game/category history with event list

        - `stats.past_esa.first_appearance` — full details of debut run

        - `stats.tenure.label` — human-readable SRC tenure string

        - `stats.communities[]` — ESA-adjacent communities found in SRC/Twitch bio


        `has_profile: false` means no runner-profile cache entry exists; trigger `POST
        /api/admin/sync/runners?slug={slug}` to populate it.'
      operationId: get_runner_profile_api_runners__slug__profile_get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runners/{slug}/pbs:
    get:
      tags:
      - runners
      summary: Get Runner Pbs
      operationId: get_runner_pbs_api_runners__slug__pbs_get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerPBsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runners/{slug}/runs:
    get:
      tags:
      - runners
      summary: Get Runner Runs
      operationId: get_runner_runs_api_runners__slug__runs_get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    RunnerSummary:
      type: object
      description: Headline statistics sourced from lineup.esamarathon.dev + speedrun.com.
      properties:
        event_count:
          type: integer
          description: Distinct ESA events attended (all-time)
        appearance_count:
          type: integer
          description: Total run appearances at ESA (all-time)
        first_event:
          anyOf:
          - type: string
          - type: 'null'
          description: Name of first ESA event, e.g. "ESA 2013"
        first_event_year:
          anyOf:
          - type: integer
          - type: 'null'
        years_active:
          anyOf:
          - type: integer
          - type: 'null'
        pb_count:
          type: integer
        pb_game_count:
          type: integer
        src_tenure_years:
          anyOf:
          - type: integer
          - type: 'null'
          description: Years on speedrun.com
        community_count:
          type: integer
        country_code:
          anyOf:
          - type: string
          - type: 'null'
        country_name:
          anyOf:
          - type: string
          - type: 'null'
          description: Self-reported; unverified
        esa_game_count:
          type: integer
          description: Distinct games run at ESA (all-time)
        esa_category_count:
          type: integer
    EsaRunEntry:
      type: object
      properties:
        game:
          type: string
        category:
          type: string
        count:
          type: integer
          description: Number of times run at ESA
        events:
          type: array
          items:
            type: string
          description: Event slugs
    PastEsaStats:
      type: object
      properties:
        event_count:
          type: integer
        appearance_count:
          type: integer
        first_event:
          anyOf:
          - type: string
          - type: 'null'
        first_event_year:
          anyOf:
          - type: integer
          - type: 'null'
        events:
          type: array
          items:
            type: string
          description: Ordered list of event names attended
        first_appearance:
          anyOf:
          - type: object
            properties:
              event_slug:
                type: string
              event_name:
                type: string
              event_start:
                type: string
                format: date-time
              scheduled:
                type: string
                format: date-time
              game:
                type: string
              category:
                type: string
              estimate:
                type: string
              stream:
                type: string
              match:
                type: string
                enum:
                - lineup
                - horaro
          - type: 'null'
        esa_runs:
          type: array
          items:
            $ref: '#/components/schemas/EsaRunEntry'
        esa_game_count:
          type: integer
        esa_category_count:
          type: integer
        sources:
          type: array
          items:
            type: string
        verified:
          type: boolean
    RunnerStats:
      type: object
      description: Detailed breakdown from runner-profile CLI (lineup + SRC). Each
        sub-key carries available/verified flags.
      properties:
        past_esa:
          anyOf:
          - $ref: '#/components/schemas/PastEsaStats'
          - type: 'null'
        tenure:
          anyOf:
          - type: object
            properties:
              available:
                type: boolean
              signup:
                anyOf:
                - type: string
                  format: date
                - type: 'null'
              years:
                anyOf:
                - type: integer
                - type: 'null'
              label:
                type: string
                description: Human-readable, e.g. "11 years on speedrun.com (since
                  2014-09-02)"
          - type: 'null'
        communities:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              source:
                type: string
              evidence:
                type: string
        country:
          anyOf:
          - type: object
            properties:
              code:
                anyOf:
                - type: string
                - type: 'null'
              name:
                anyOf:
                - type: string
                - type: 'null'
              source:
                type: string
              confidence:
                type: string
          - type: 'null'
        personal_bests:
          anyOf:
          - type: object
            properties:
              game_count:
                type: integer
              category_count:
                type: integer
              top_games:
                type: array
                items:
                  type: string
          - type: 'null'
    RunnerProfileResponse:
      type: object
      description: 'Composite runner profile returned by GET /api/runners/{slug}/profile.


        This is the authoritative source for **all historical ESA data**: event counts,
        run history, first appearance, ESA games, country, community, and SRC tenure.
        The run_count/esa_count fields on RunnerDTO (GET /api/runners/{slug}) are
        scoped to the **current marathon only** — use this endpoint for all-time figures.'
      required:
      - slug
      - display_name
      - twitch
      - discord
      - twitter
      - pronouns
      - sources
      - errors
      - has_profile
      properties:
        slug:
          type: string
        display_name:
          type: string
        twitch:
          type: string
        discord:
          type: string
        twitter:
          type: string
        pronouns:
          type: string
        pronunciation:
          type: string
        summary:
          anyOf:
          - $ref: '#/components/schemas/RunnerSummary'
          - type: 'null'
        stats:
          anyOf:
          - $ref: '#/components/schemas/RunnerStats'
          - type: 'null'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/BriefSidecarSource'
        errors:
          type: array
          items:
            type: string
        has_profile:
          type: boolean
          description: False when no runner-profile cache entry exists for this runner
    RunnerDTOResponse:
      type: object
      description: 'Runner identity and **current-marathon** run summary.


        **run_count** and **esa_count** reflect only runs in this event''s schedule.
        For all-time ESA history (17 events, 26 runs, first year, etc.) call GET /api/runners/{slug}/profile
        which reads from the runner-profile cache.'
      required:
      - slug
      - display_name
      - twitch
      - discord
      - twitter
      - pronouns
      properties:
        slug:
          type: string
        display_name:
          type: string
        twitch:
          type: string
        discord:
          type: string
        twitter:
          type: string
        pronouns:
          type: string
        pronunciation:
          type: string
        run_count:
          type: integer
          description: Runs in the **current marathon** only
        esa_count:
          type: integer
          description: Distinct streams in the **current marathon** only. For all-time
            event count see /profile
        first_esa:
          anyOf:
          - type: string
          - type: 'null'
          description: Date of first scheduled run in the current marathon (not historical
            debut)
        upcoming_runs:
          type: array
          items:
            $ref: '#/components/schemas/RunDTO'
          description: Next 5 upcoming runs for this runner
    RunnerPBsResponse:
      type: object
      required:
      - slug
      - display_name
      - pbs
      - has_pbs
      properties:
        slug:
          type: string
        display_name:
          type: string
        pbs:
          type: array
          items:
            $ref: '#/components/schemas/RunnerPBEntry'
        has_pbs:
          type: boolean
  securitySchemes:
    hostSession:
      type: apiKey
      in: cookie
      name: esa_session
      description: Set via POST /api/login. Valid 24h.
    adminSession:
      type: apiKey
      in: cookie
      name: esa_admin_session
      description: Set via POST /api/admin/login. Valid 1h.
servers:
- url: https://esa-hosting.codescales.xyz
  description: Production
...